Create initial working version
This version was created based on Boxtron source code (v0.5.0) with all
DOSBox-specific features removed and basically no tests. Python modules
are named the same way as in Boxtron, but any non-essential function
was removed; functions that are likely going to be useful in future were
preserved.

This initial version has serious limitations (e.g. does not work at all
with Steam games using Sierra Launcher), but is already able to start
some games (tested with Larry 6 VGA and Larry 7).